Safari-611.1.10/LayoutTests/js/dom/document-all-is-callable-builtins-expected.txt

document.all is callable (builtins)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testArray() is true
PASS testTypedArray() is true
PASS testRegExp() is true
PASS testOtherBuiltins() is true
PASS successfullyParsed is true

TEST COMPLETE
